fix: address PR #87 AI review findings
Blocking fix:
- settings_routes: apply port normalization to fallback HUGO_SERVER_BASE_URL
  value before assigning to hugo_manager.server_url

Architecture improvements:
- Move /api/frontmatter/generate from image_routes to ai_routes (separate
  fm_bp Blueprint) — keeps AI concerns together
- Extract /api/references/* and /api/posts/search from post_routes into
  new references_routes.py — separates reference domain
- Move /api/version from settings_routes to page_routes — version is not
  a setting

Quality cleanup:
- Remove unused _stop_on_error variable from publish_routes bulk_publish

All 96 tests pass. Ruff lint clean.
