+++ routes/post_routes.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 """
-文章管理、缓存、引用关系相关路由
+文章管理、缓存相关路由
 """
 
 import logging
@@ -70,35 +70,5 @@ def register_post_routes(registry):
             return jsonify({"success": True, "stats": stats})
         else:
             return jsonify({"success": False, "message": "缓存未启用"}), 400
-
-    @bp.route("/api/references/scan", methods=["POST"])
-    def scan_references():
-        """扫描所有文章的引用关系"""
-        try:
-            registry.ref_service.scan_all()
-            refs = registry.ref_service.db.get_all_references()
-            return jsonify({"success": True, "references": refs})
-        except Exception as e:
-            return jsonify({"success": False, "message": str(e)}), 500
-
-    @bp.route("/api/references/backlinks")
-    def get_backlinks():
-        """获取反向链接"""
-        file_path = request.args.get("path")
-        if not file_path:
-            return jsonify({"success": False, "message": "缺少 path 参数"}), 400
-
-        backlinks = registry.ref_service.get_backlinks(file_path)
-        return jsonify({"success": True, "backlinks": backlinks})
-
-    @bp.route("/api/posts/search")
-    def search_posts():
-        """文章搜索（自动补全用）"""
-        query = request.args.get("q", "")
-        if not query:
-            return jsonify({"success": True, "posts": []})
-
-        posts = registry.ref_service.search_posts(query)
-        return jsonify({"success": True, "posts": posts})
 
     return bp
